+++ src/main/java/org/springframework/gresur/repository/ConfiguracionRepository.java
@@ -3,6 +3,6 @@ package org.springframework.gresur.repository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.gresur.model.Configuracion;
 
-public interface ConfiguracionRepository extends CrudRepository<Configuracion, Integer>{
+public interface ConfiguracionRepository extends CrudRepository<Configuracion, Long>{
 	
 }
