+++ src/main/java/org/springframework/gresur/repository/PedidoRepository.java
@@ -3,6 +3,6 @@ package org.springframework.gresur.repository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.gresur.model.Pedido;
 
-public interface PedidoRepository extends CrudRepository<Pedido, Integer>{
+public interface PedidoRepository extends CrudRepository<Pedido, Long>{
 
 }
