+++ src/main/java/org/springframework/gresur/repository/ProveedorRepository.java
@@ -4,7 +4,7 @@ import org.springframework.dao.DataAccessException;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.gresur.model.Proveedor;
 
-public interface ProveedorRepository extends CrudRepository<Proveedor, Integer>{
+public interface ProveedorRepository extends CrudRepository<Proveedor, Long>{
 	
 	Proveedor findByNIF (String NIF) throws DataAccessException;
